# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Psychological Monitoring Platform",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patients/me": {
      "get": {
        "tags": [
          "Patient Profile"
        ],
        "summary": "Get My Profile",
        "operationId": "get_my_profile_api_v1_patients_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Patient Profile"
        ],
        "summary": "Update My Profile",
        "operationId": "update_my_profile_api_v1_patients_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/patients/me/conditions": {
      "get": {
        "tags": [
          "Patient Profile"
        ],
        "summary": "List My Conditions",
        "operationId": "list_my_conditions_api_v1_patients_me_conditions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PatientConditionOut"
                  },
                  "type": "array",
                  "title": "Response List My Conditions Api V1 Patients Me Conditions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Patient Profile"
        ],
        "summary": "Add My Condition",
        "operationId": "add_my_condition_api_v1_patients_me_conditions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientConditionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientConditionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/patients/me/conditions/{condition_id}": {
      "delete": {
        "tags": [
          "Patient Profile"
        ],
        "summary": "Remove My Condition",
        "operationId": "remove_my_condition_api_v1_patients_me_conditions__condition_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "condition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Condition Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mood-entries": {
      "post": {
        "tags": [
          "Mood Tracking"
        ],
        "summary": "Create Mood Entry",
        "operationId": "create_mood_entry_api_v1_mood_entries_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoodEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoodEntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Mood Tracking"
        ],
        "summary": "List Mood Entries",
        "operationId": "list_mood_entries_api_v1_mood_entries_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoodEntryOut"
                  },
                  "title": "Response List Mood Entries Api V1 Mood Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mood-entries/trend": {
      "get": {
        "tags": [
          "Mood Tracking"
        ],
        "summary": "Mood Trend",
        "operationId": "mood_trend_api_v1_mood_entries_trend_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoodTrendPoint"
                  },
                  "title": "Response Mood Trend Api V1 Mood Entries Trend Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "type": "array",
                  "title": "Response List Notifications Api V1 Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark Notification Read",
        "operationId": "mark_notification_read_api_v1_notifications__notification_id__read_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/register": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Register Device",
        "description": "Registers an FCM device token for push notifications.\n\nNOTE (future step): persist the token (e.g., in a `user_devices` table)\nand wire up the FCM sending service. For Step 2 this endpoint validates\nthe request shape and is ready for that integration.",
        "operationId": "register_device_api_v1_devices_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityLevel": {
        "type": "string",
        "enum": [
          "sedentary",
          "light",
          "moderate",
          "active"
        ],
        "title": "ActivityLevel"
      },
...
      "ChronicConditionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "description_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Ar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name_en",
          "name_ar"
        ],
        "title": "ChronicConditionOut"
      },
...
      "DeviceRegisterRequest": {
        "properties": {
          "fcm_token": {
            "type": "string",
            "title": "Fcm Token"
          }
        },
        "type": "object",
        "required": [
          "fcm_token"
        ],
        "title": "DeviceRegisterRequest"
      },
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other"
        ],
        "title": "Gender"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoodEntryCreate": {
        "properties": {
          "mood_value": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Mood Value"
          },
          "note_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Ar"
          }
        },
        "type": "object",
        "required": [
          "mood_value"
        ],
        "title": "MoodEntryCreate"
      },
      "MoodEntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "mood_value": {
            "type": "integer",
            "title": "Mood Value"
          },
          "note_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Ar"
          },
          "source": {
            "$ref": "#/components/schemas/MoodSource"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mood_value",
          "source",
          "recorded_at"
        ],
        "title": "MoodEntryOut"
      },
      "MoodSource": {
        "type": "string",
        "enum": [
          "manual",
          "interview_derived"
        ],
        "title": "MoodSource"
      },
      "MoodTrendPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "average_mood": {
            "type": "number",
            "title": "Average Mood"
          }
        },
        "type": "object",
        "required": [
          "date",
          "average_mood"
        ],
        "title": "MoodTrendPoint"
      },
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "title_ar": {
            "type": "string",
            "title": "Title Ar"
          },
          "body_ar": {
            "type": "string",
            "title": "Body Ar"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title_ar",
          "body_ar",
          "is_read",
          "status",
          "created_at"
        ],
        "title": "NotificationOut"
      },
      "NotificationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "sent",
          "failed"
        ],
        "title": "NotificationStatus"
      },
      "NotificationType": {
        "type": "string",
        "enum": [
          "daily_checkin",
          "follow_up",
          "mood_reminder",
          "recommendation_alert",
          "engagement",
          "risk_alert_admin"
        ],
        "title": "NotificationType"
      },
      "PatientConditionCreate": {
        "properties": {
          "chronic_condition_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chronic Condition Id"
          },
          "diagnosed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosed At"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "chronic_condition_id"
        ],
        "title": "PatientConditionCreate"
      },
      "PatientConditionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chronic_condition": {
            "$ref": "#/components/schemas/ChronicConditionOut"
          },
          "diagnosed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosed At"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chronic_condition",
          "is_primary"
        ],
        "title": "PatientConditionOut"
      },
      "PatientProfileOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "disease_duration_months": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Duration Months"
          },
          "medications": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medications"
          },
          "sleep_hours_avg": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sleep Hours Avg"
          },
          "activity_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "social_support_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SocialSupportLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "medical_background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Background"
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/PatientConditionOut"
            },
            "type": "array",
            "title": "Conditions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "onboarding_completed",
          "created_at"
        ],
        "title": "PatientProfileOut"
      },
      "PatientProfileUpdate": {
        "properties": {
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "disease_duration_months": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Duration Months"
          },
          "medications": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medications"
          },
          "sleep_hours_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sleep Hours Avg"
          },
          "activity_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "social_support_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SocialSupportLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "medical_background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Background"
          }
        },
        "type": "object",
        "title": "PatientProfileUpdate"
      },
...
      "SocialSupportLevel": {
        "type": "string",
        "enum": [
          "none",
          "low",
          "medium",
          "high"
        ],
        "title": "SocialSupportLevel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}